+++ CHANGES.md
@@ -1,6 +1,6 @@
 # 2.10.0
 
-* Added `Relation.active`, and excluded inactive relations from `Model`.relations
+* Added `Relation.active`, and excluded inactive relations from `Model.relations`
 
 # 2.9.0